+++ viewlegacy/src/androidMain/kotlin/tech/skot/core/components/inputs/SKInputWithSuggestionsView.kt
@@ -25,6 +25,14 @@ class SKInputWithSuggestionsView(
 
     private var watching: TextWatcher? = null
 
+    fun onOnFocusChange(onFocusChange: ((hasFocus:Boolean) -> Unit)?) {
+        if (onFocusChange != null) {
+            autoComplete.setOnFocusChangeListener { _, hasFocus ->
+                onFocusChange.invoke(hasFocus)
+            }
+        }
+    }
+
     fun onOnInputText(onInputText: (String?) -> Unit) {
         val watcher =
             object : TextWatcher {