+++ viewmodel/src/commonMain/kotlin/tech/skot/core/components/inputs/SKInputWithSuggestions.kt
@@ -3,7 +3,8 @@ package tech.skot.core.components.inputs
 import tech.skot.core.di.coreViewInjector
 import tech.skot.core.view.Color
 
-class SKInputWithSuggestions<D : Any?>(
+
+open class SKInputWithSuggestions<D : Any?>(
     hint: String? = null,
     errorInitial: String? = null,
     initialChoices: List<D> = emptyList(),
@@ -34,5 +35,19 @@ class SKInputWithSuggestions<D : Any?>(
             onInputText = onInputText,
             dropDownDisplayedInitial = false,
             oldSchoolModeHint = oldSchoolModeHint,
-        )
+            onFocusChange = {
+                if (it) {
+                    onFocus()
+                } else {
+                    onFocusLost()
+                }
+            },
+    )
+
+    protected open fun onFocus() {
+    }
+
+    protected open fun onFocusLost() {
+
+    }
 }