+++ viewmodelTests/src/main/kotlin/tech/skot/core/di/CoreViewInjectorMock.kt
@@ -215,6 +215,7 @@ class CoreViewInjectorMock : CoreViewInjector {
         hiddenInitial: Boolean?,
         dropDownDisplayedInitial: Boolean,
         onInputText: (input: String?) -> Unit,
+        onFocusChange: ((hasFocus:Boolean) -> Unit)?,
         oldSchoolModeHint: Boolean,
     ): SKInputWithSuggestionsVC {
         return SKInputWithSuggestionsViewMock(
@@ -227,6 +228,7 @@ class CoreViewInjectorMock : CoreViewInjector {
             hiddenInitial,
             dropDownDisplayedInitial,
             onInputText,
+            onFocusChange,
             oldSchoolModeHint,
         )
     }